requirements.txt
python-chess>=1.10.0\npytest>=7.0.0
